+++ app/Http/Controllers/Backend/CategoriaController.php
@@ -77,7 +77,7 @@ class CategoriaController extends AppBaseController
         $categoria = $this->categoriaRepository->findWithoutFail($id);
 
         if (empty($categoria)) {
-            Flash::error('Categoria no encontrado');
+            Flash::error('Categoria no encontrada');
 
             return redirect(route('backend.categorias.index'));
         }
@@ -97,7 +97,7 @@ class CategoriaController extends AppBaseController
         $categoria = $this->categoriaRepository->findWithoutFail($id);
 
         if (empty($categoria)) {
-            Flash::error('Categoria no encontrado');
+            Flash::error('Categoria no encontrada');
 
             return redirect(route('backend.categorias.index'));
         }
@@ -118,7 +118,7 @@ class CategoriaController extends AppBaseController
         $categoria = $this->categoriaRepository->findWithoutFail($id);
 
         if (empty($categoria)) {
-            Flash::error('Categoria no encontrado');
+            Flash::error('Categoria no encontrada');
 
             return redirect(route('backend.categorias.index'));
         }
@@ -142,11 +142,16 @@ class CategoriaController extends AppBaseController
         $categoria = $this->categoriaRepository->findWithoutFail($id);
 
         if (empty($categoria)) {
-            Flash::error('Categoria no encontrado');
+            Flash::error('Categoria no encontrada');
 
             return redirect(route('backend.categorias.index'));
         }
+        $productos = $categoria->productos()->get();
+        if (!empty($productos)){
+            Flash::error('No es posible eliminar la Categoria dado que tiene Productos asociados');
 
+            return redirect(route('backend.categorias.index'));
+        }
         $this->categoriaRepository->delete($id);
 
         Flash::success('Categoria borrada exitosamente');
